# Speedup STAC collections endpoint