As reported in #802, we were experiencing timeouts or extremely slow responses to the `/stac/collections` endpoint. The implementation was doing a lot of extra data transfer and work, by transferring full geometries from PostGIS into Python before simplifying them down to a bbox.

This moves the processing up into PostGIS, and in my testing gives a speedup between 10x and 121x, and potentially more in different environments. I only tested on my laptop.

Running with local PostGIS server, M4 MacBook Pro.

## pytest-benchmark
Loading `/stac/collections` with the data loaded in `test_stac.py`

<img width="1841" height="114" alt="image" src="https://github.com/user-attachments/assets/bdf7f158-baa0-4dcd-92e9-52210cc69f0a" />

Small dataset, but about 10x faster.

## Running datacube-explorer locally
Using a copy of our database with the horrendously complex geometries

Run the server:
```
uv run gunicorn -w 3 --threads 2 -k gthread --config python:cubedash.gunicorn_config 'cubedash:create_app()'
```

Run benchmark with [hey](https://github.com/rakyll/hey)

Before this change:
```
# 2 workers, 4 requests
$ hey -c 2 -n 4 -t 30 http://127.0.0.1:8000/stac/collections

Summary:
  Total:        23.5620 secs
  Slowest:      11.9025 secs
  Fastest:      11.6424 secs
  Average:      11.7767 secs
  Requests/sec: 0.1698
```

After this change:
```
# For fast, defaults to 50 workers and 200 requests
$ hey -t 30 http://127.0.0.1:8000/stac/collections

Summary:
  Total:        9.8161 secs
  Slowest:      2.6334 secs
  Fastest:      0.2301 secs
  Average:      1.8218 secs
  Requests/sec: 20.3746
```

121x faster.

I haven't added any request caching discussed in the issue, that can be done separately.

Fixes #802 

<!-- readthedocs-preview datacube-explorer start -->
----
📚 Documentation preview 📚: https://datacube-explorer--871.org.readthedocs.build/en/871/

<!-- readthedocs-preview datacube-explorer end -->